nethra_original_backup/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database dependencies
aiosqlite==0.19.0

# Security dependencies
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# AI/ML dependencies
tensorflow==2.15.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Async and concurrency
aiofiles==23.2.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
marshmallow==3.20.1
jsonschema==4.19.2

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Monitoring and performance
psutil==5.9.6
prometheus-client==0.19.0

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1